+++ model/schema_combo_test.go
@@ -6,4 +6,5 @@ func TestCombo(t *testing.T) {
 	// TODO: Schema.GetOrCreate
 	// TODO: Schema.GetOrCreateTx
 	// TODO: Schema.GetAndLockTx
+	// TODO: Schema.GetAndLockTxItems
 }